introduce a change in the semantics of integers
The Oberon+ semantics says that integer literals must be assigned a
token type that matches its bit widths. That is, values between 0 and
255 are BYTES, values between -32768 and 32767 are INT16. Unless, the
user specifies that the type of integer by appending x/X [char], l/L
[INT64] or i/I [INT32].

This is causing some type problems during IR code generation where
a variable defined as 'integer' become BYTE per the value assigned
to it, say '4'.

So this change replaces this rule with one where the assigned type
take precedence. For integer literals, they get a default type if INT